fix 表格bug

+++ app/src/main/java/com/zip/zipandroid/adapter/ZipRepaymentPlanAdapter.kt
@@ -19,7 +19,7 @@ class ZipRepaymentPlanAdapter : BaseQuickAdapter<ZipRepayment, BaseViewHolder>(R
         val plan_bottom_line_view = holder.getView<VerticalDashedLineView>(R.id.plan_bottom_line_view)
         item_zip_repayment_count_tv.setText((holder.adapterPosition + 1).toString() + "/" + data.size.toString())
         item_zip_repayment_time_tv.setText(ZipTimeUtils.formatTimestampToDate(item.shouldTime))
-        item_payment_price_tv.setText(item.shouldAmount.toN())
+        item_payment_price_tv.setText(item.shouldAmount.toDouble().toN())
         plan_top_line_view.visible = holder.adapterPosition != 0
         plan_bottom_line_view.visible = holder.adapterPosition != data.size - 1
     }
